Modified forced fetch and fixed on-page scripts' fetch
The "random method" has been replaced with the "date method" to make 
cache busting working.
Every time the a page is reloaded or opened first (not fetched from an 
already opened page), the current time will be get and used as version 
for all the files fetched until the page is reloaded or closed and 
opened again in a new session.

Scripts written directly to the HTML page will be correctly fetched now.